Use non-wakeup alarm for delay constraint.
The delay alarm does not need to go off exactly when the time has
passed, so a non-wakeup alarm is suitable for the delay constraint.
A job is expected to go off as soon as its deadline constraint is
satisfied, so we leave the deadline alarm as a wakeup alarm.

Bug: 138465403
Bug: 147162283
Test: atest com.android.server.job.controllers.TimeControllerTest
Test: atest CtsJobSchedulerTestCases
Change-Id: I2a38c0dfc9a1efd862e3a1a4c8351dd4b522021e
Merged-In: I5756bd805c68dea51b3a5aa74419b195407e1a30
Signed-off-by: Sebita <kjjjnob.seba@gmail.com>